+++ database/migrations/tenant/2019_07_11_080725_add_device_id_field_to_trailer_to_trailer_boxes_table.php
@@ -15,7 +15,7 @@ class AddDeviceIdFieldToTrailerToTrailerBoxesTable extends Migration
     {
         if (!Schema::hasColumn('trailer_boxes','device_id')){
             Schema::table('trailer_boxes', function (Blueprint $table) {
-                $table->unsignedInteger('device_id')->nullable();
+                $table->unsignedBigInteger('device_id')->nullable();
                 $table->foreign('device_id')->references('id')->on('devices');
             });
         }
